+++ test/libsolidity/syntaxTests/abiEncoder/external_functions_taking_internal_types_struct_with_mapping.sol
@@ -6,4 +6,3 @@ contract C {
 }
 // ----
 // TypeError 4103: (105-113): Types containing (nested) mappings can only be parameters or return variables of internal or library functions.
-// TypeError 4061: (105-113): Type struct C.S is only valid in storage because it contains a (nested) mapping.