+++ test/libsolidity/syntaxTests/functionCalls/calldata_struct_array_argument_with_internal_data_type_inside.sol
@@ -1,0 +1,8 @@
+contract C {
+	struct S {
+		function() a;
+	}
+	function f(S[2] calldata) public {}
+}
+// ----
+// TypeError 4103: (56-69): Internal type is not allowed for public or external functions.