+++ test/libsolidity/syntaxTests/types/mapping/mapping_struct_data_location_memory.sol
@@ -5,4 +5,3 @@ contract C {
 }
 // ----
 // TypeError 4103: (104-112): Types containing (nested) mappings can only be parameters or return variables of internal or library functions.
-// TypeError 4061: (104-112): Type struct C.S is only valid in storage because it contains a (nested) mapping.