+++ Changelog.md
@@ -9,6 +9,8 @@ Compiler Features:
 Bugfixes:
  * Code Generator: Fix a crash when using ``@use-src`` and compiling from Yul to ewasm.
  * SMTChecker: Fix internal error when an unsafe target is solved more than once and the counterexample messages are different.
+ * Fix internal error when a function has a calldata struct argument with an internal type inside.
+
 
 
 ### 0.8.10 (2021-11-09)